refactor(core): extraer helpers compartidos de admin pages
- Crear admin/_shared/helpers.tsx con formatDate, formatDateTime,
  roleBadgeClass e InfoRow compartidos entre tenant detail y user detail
- Crear tenants/[id]/_helpers.tsx con constantes, badge helpers,
  FeatureFlag, PhaseStepper, PhaseBadge y AdminTenantDetailSkeleton
- Crear users/[id]/_helpers.tsx con PROVIDER_LABELS/ICONS, statusBadgeClass,
  fullName, AuthCardEmptyState y AdminUserDetailSkeleton
- Reducir tenants/[id]/page.tsx de ~1615 a 1275 lineas (-340)
- Reducir users/[id]/page.tsx de ~1195 a 1033 lineas (-162)
- Eliminar duplicacion de helpers entre ambas pages (DRY)

Refs: #244
